complete email module and migration

+++ PostModule/PostModule.Application.Contract/UserPostApplication/Command/CreatePostOrder.cs
@@ -2,14 +2,14 @@
 
 public class CreatePostOrder
 {
-    public CreatePostOrder(int userId, int packageId,int price)
+    public CreatePostOrder(long userId, int packageId,int price)
     {
         UserId = userId;
         PackageId = packageId;
         Price = price;
     }
 
-    public int UserId { get; private set; }
+    public long UserId { get; private set; }
     public int PackageId { get; private set; }
     public int Price { get; private set; }
 }
