+++ PostModule/PostModule.Domain/UserPostAgg/IPOstOrderRepository.cs
@@ -6,7 +6,7 @@ namespace PostModule.Domain.UserPostAgg
 {
     public interface IPOstOrderRepository : IRepository<int,PostOrder>
     {
-        Task<PostOrderUserPanelModel> GetPostOrderNotPaymentForUser(int userId);
-        Task<PostOrder> GetPostOrderNotPaymentForUserAsync(int userId);
+        Task<PostOrderUserPanelModel> GetPostOrderNotPaymentForUser(long userId);
+        Task<PostOrder> GetPostOrderNotPaymentForUserAsync(long userId);
     }
 }
