+++ PostModule/PostModule.Domain/UserPostAgg/IUserPostRepository.cs
@@ -5,7 +5,7 @@ namespace PostModule.Domain.UserPostAgg
 {
     public interface IUserPostRepository : IRepository<int,UserPost>
     {
-        Task<UserPost> GetForUser(int userId);
+        Task<UserPost> GetForUser(long userId);
         Task<UserPost> GetByApiCode(string apiCode);
     }
 }
