+++ PostModule/PostModule.Infrastracture.EF/Repositories/PackageRepository.cs
@@ -13,7 +13,7 @@ internal class PackageRepository : Repository<int, Package>, IPackageRepository
         _context = context;
     }
 
-    public async Task<CreatePostOrder> GetCreatePostModelAsync(int userId, int packageId)
+    public async Task<CreatePostOrder> GetCreatePostModelAsync(long userId, int packageId)
     {
         var package = await _context.Packages.FindAsync(packageId);
         return new CreatePostOrder(userId,package.Id,package.Price);
